Improve aspirant's register/login system.
Now when an aspirants login into the system, a fund is created with its corresponding status.

If there aren't funds already, this new fund has the 'gathering funds' status.
In othercase, it is saved with 'waiting' status.

Also, when a duplicate email is registerd, systems execute a find search instead a regist.

Both behaviors returns both objects, aspirant and its fund data

Co-authored-by: Feli <srgatoiscool@gmail.com>